weather_bot/requirements.txt
aiogram==3.18.0
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.11.0
matplotlib==3.10.1
pandas==2.2.3
numpy==2.2.3
